mosque project esp32s codes in .ino format
Before writing the code, it is mandatory to download the necessary libraries in the code to the arudino ide program, otherwise the code will not work. Also, in order to upload to the esp32s card, you must introduce the esp32 cards to the arduino ide. If you ask this to chatgpt "What should I do to upload code from the arduino ide program to the esp32s card?", they will help you with love...